refactor: 잘못된 테스트 클래스명 수정
